Other: Renamed the BalloonToolbarEditor file to balloontoolbareditor.js to match the naming convention. Closes #5.
