+++ tests/manual/balloontoolbareditor.html
