+++ tests/manual/balloontoolbareditor.md
